[Rename] 게시물 컴포넌트 위치 수정
