+++ src/components/home/post/PostList.js
@@ -1,0 +1,14 @@
+import StyledDiv from '@styles/home/post/PostList-styled';
+import PostItem from './PostItem';
+
+const PostList = ({ posts }) => {
+  return (
+    <StyledDiv>
+      {posts.map((post) => (
+        <PostItem key={post.id} post={post} />
+      ))}
+    </StyledDiv>
+  );
+};
+
+export default PostList;
